Improvements to the edit view
Lon/Lat now default to 0 if left blank, alt will default to null. All show placeholders of 0.
The user can delete any position, including the first one if there is at least one listed.
Swipeview is disabled when editing a point

+++ Groundsman/Services/FeatureStore.cs
@@ -229,7 +229,7 @@ namespace Groundsman.Services
             {
                 try
                 {
-                    double metadataFloatValue = Convert.ToSingle(feature.Properties["metadataFloatValue"]);
+                    float metadataFloatValue = Convert.ToSingle(feature.Properties["metadataFloatValue"]);
                     feature.Properties["metadataFloatValue"] = metadataFloatValue;
                 }
                 catch
